+++ src/Network/Network/ChangeLog.md
@@ -55,6 +55,7 @@
 * Update cmdlets to return connections in Network Virtual Appliance
     -`Network Virtual Appliance`
 * Allowed not to provide `Rules` in `PSApplicationGatewayFirewallPolicyManagedRuleGroupOverride`, which would return an empty `RuleID` to be passed to NRP.
+* Add optional parameter 'AdminState' to Express Route Virtual Network Gateway
 
 ## Version 5.7.0
 * Onboarded `Microsoft.HardwareSecurityModules/cloudHsmClusters` to private link cmdlets